refactor(gateway): relocate SSE writer to data-plane/shared
writeSSEFrames is consumed by both chat/*/respond.ts (via respondSseStream) and
data-plane/shared/passthrough-serve.ts, but was living under chat/shared/stream/
so passthrough reached across the boundary with `../chat/shared/stream/sse.ts`.
Moves the stream/ subdirectory up one level so both consumer families import
their local peer instead. chat/shared/ now holds only chat-scoped utilities;
utilities shared across chat and passthrough live under data-plane/shared/.
