+++ packages/gateway/src/data-plane/shared/stream/sse.ts
