+++ packages/gateway/src/data-plane/shared/stream/sse_test.ts
@@ -3,7 +3,7 @@ import { streamSSE } from 'hono/streaming';
 import { test } from 'vitest';
 
 import { writeSSEFrames } from './sse.ts';
-import { FakeTime } from '../../../../test-time.ts';
+import { FakeTime } from '../../../test-time.ts';
 import { parseSSEStream } from '@floway-dev/protocols/common';
 import { sseCommentFrame, type SseFrame, sseFrame } from '@floway-dev/protocols/common';
 import { assertEquals } from '@floway-dev/test-utils';
